% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trapz.R
\name{trapz.density}
\alias{trapz.density}
\alias{cumtrapz.density}
\title{trapz and cumtrapz}
\usage{
trapz.density(x)

cumtrapz.density(x)
}
\arguments{
\item{x}{returned value of function \link[stats]{density}}
}
\value{
Function \code{\link[=trapz.density]{trapz.density()}} returns a \link[base]{numeric} scalar.

Function \code{\link[=cumtrapz.density]{cumtrapz.density()}} returns a \link[base]{numeric} \link[base]{matrix}.
}
\description{
..
}
\details{
..
}
\examples{
rnorm(1e2) |> density() |> trapz.density()

rnorm(1e2) |> density(n = 16L) |> cumtrapz.density()
}
\keyword{internal}
